Optimize ArrayList allocations
Motivation:

There are tons of places where jackson-jq allocates ArrayLists with default size (which is 10) while the target size is known. In some cases, we even already know that the target is a singleton, hence no need to allocate an ArrayList.

This is a waste of allocations.

Modifications:

Introduce fast paths in `FieldAccess`, `ResolvedAllFieldAccess`, `ResolvedIndexFieldAccess` and `ResolvedStringFieldAccess` for cases  where target collection is a singleton.

Result:

Less allocations, +20% throughput improvement on some of my internal benchmarks, eg when extracting one single field.