core: remove warning regarding pattern matches in path
This was not right. If I understand correctly, what jq does not support
is getting a path of $variables. I misunderstand as getting a path from
a singlely pattern matched variable (e.g. `.foo as $a | path($a)`) did
work in jq-1.5 (not in jq-1.6). I will disable getting a path from
$variable altogether in the future commit.